Code review update: Split SA1402SettingsConfiguration.cs in two files.
