+++ StyleCop.Analyzers/StyleCop.Analyzers.Test/MaintainabilityRules/SA1402SettingsConfiguration.cs
@@ -3,10 +3,6 @@
 
 namespace StyleCop.Analyzers.Test.MaintainabilityRules
 {
-    using System.Collections.Generic;
-    using System.Diagnostics.CodeAnalysis;
-    using System.Linq;
-
     public enum SA1402SettingsConfiguration
     {
         /// <summary>
@@ -23,36 +19,5 @@ namespace StyleCop.Analyzers.Test.MaintainabilityRules
         /// Provide custom settings that configure the tested type as not being a top level type
         /// </summary>
         ConfigureAsNonTopLevelType,
-    }
-
-    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1649:File name must match first type name.", Justification = "Extension method for enum above.")]
-    public static class SA1402SettingsConfigurationExtensions
-    {
-        public static string GetSettings(this SA1402SettingsConfiguration configuration, string keyword)
-        {
-            if (configuration == SA1402SettingsConfiguration.KeepDefaultConfiguration)
-            {
-                return null;
-            }
-
-            var keywords = new List<string> { "class", "interface", "struct", "enum", "delegate" };
-            if (configuration == SA1402SettingsConfiguration.ConfigureAsNonTopLevelType)
-            {
-                keywords.Remove(keyword);
-            }
-
-            var keywordsStr = string.Join(", ", keywords.Select(x => "\"" + x + "\""));
-
-            var settings = $@"
-{{
-  ""settings"": {{
-    ""maintainabilityRules"": {{
-      ""topLevelTypes"": [{keywordsStr}]
-    }}
-  }}
-}}";
-
-            return settings;
-        }
     }
 }
